+++ front/src/Api/iframe/room.ts
@@ -1,28 +1,14 @@
-import { Subject } from "rxjs";
+import { Observable, Subject } from "rxjs";
 
-import { isDataLayerEvent } from "../Events/DataLayerEvent";
 import { EnterLeaveEvent, isEnterLeaveEvent } from "../Events/EnterLeaveEvent";
-import { isGameStateEvent } from "../Events/GameStateEvent";
 
 import { IframeApiContribution, queryWorkadventure, sendToWorkadventure } from "./IframeApiContribution";
 import { apiCallback } from "./registeredCallbacks";
 
 import type { ITiledMap } from "../../Phaser/Map/ITiledMap";
-import type { DataLayerEvent } from "../Events/DataLayerEvent";
-import type { GameStateEvent } from "../Events/GameStateEvent";
 
 const enterStreams: Map<string, Subject<EnterLeaveEvent>> = new Map<string, Subject<EnterLeaveEvent>>();
 const leaveStreams: Map<string, Subject<EnterLeaveEvent>> = new Map<string, Subject<EnterLeaveEvent>>();
-const dataLayerResolver = new Subject<DataLayerEvent>();
-
-let immutableDataPromise: Promise<GameStateEvent> | undefined = undefined;
-
-interface Room {
-    id: string;
-    mapUrl: string;
-    map: ITiledMap;
-    startLayer: string | null;
-}
 
 interface TileDescriptor {
     x: number;
@@ -31,19 +17,17 @@ interface TileDescriptor {
     layer: string;
 }
 
-export function getGameState(): Promise<GameStateEvent> {
-    if (immutableDataPromise === undefined) {
-        immutableDataPromise = queryWorkadventure({ type: "getState", data: undefined });
-    }
-    return immutableDataPromise;
-}
+let roomId: string | undefined;
 
-function getDataLayer(): Promise<DataLayerEvent> {
-    return new Promise<DataLayerEvent>((resolver, thrower) => {
-        dataLayerResolver.subscribe(resolver);
-        sendToWorkadventure({ type: "getDataLayer", data: null });
-    });
-}
+export const setRoomId = (id: string) => {
+    roomId = id;
+};
+
+let mapURL: string | undefined;
+
+export const setMapURL = (url: string) => {
+    mapURL = url;
+};
 
 export class WorkadventureRoomCommands extends IframeApiContribution<WorkadventureRoomCommands> {
     callbacks = [
@@ -59,13 +43,6 @@ export class WorkadventureRoomCommands extends IframeApiContribution<Workadventu
             typeChecker: isEnterLeaveEvent,
             callback: (payloadData) => {
                 leaveStreams.get(payloadData.name)?.next();
-            },
-        }),
-        apiCallback({
-            type: "dataLayer",
-            typeChecker: isDataLayerEvent,
-            callback: (payloadData) => {
-                dataLayerResolver.next(payloadData);
             },
         }),
     ];
@@ -102,23 +79,31 @@ export class WorkadventureRoomCommands extends IframeApiContribution<Workadventu
             },
         });
     }
-    getCurrentRoom(): Promise<Room> {
-        return getGameState().then((gameState) => {
-            return getDataLayer().then((mapJson) => {
-                return {
-                    id: gameState.roomId,
-                    map: mapJson.data as ITiledMap,
-                    mapUrl: gameState.mapUrl,
-                    startLayer: gameState.startLayerName,
-                };
-            });
-        });
+    async getTiledMap(): Promise<ITiledMap> {
+        const event = await queryWorkadventure({ type: "getMapData", data: undefined });
+        return event.data as ITiledMap;
     }
     setTiles(tiles: TileDescriptor[]) {
         sendToWorkadventure({
             type: "setTiles",
             data: tiles,
         });
+    }
+
+    get id(): string {
+        if (roomId === undefined) {
+            throw new Error("Room id not initialized yet. You should call WA.room.id within a WA.onInit callback.");
+        }
+        return roomId;
+    }
+
+    get mapURL(): string {
+        if (mapURL === undefined) {
+            throw new Error(
+                "mapURL is not initialized yet. You should call WA.room.mapURL within a WA.onInit callback."
+            );
+        }
+        return mapURL;
     }
 }