+++ pusher/src/Controller/IoSocketController.ts
@@ -17,6 +17,7 @@ import {
     ServerToClientMessage,
     CompanionMessage,
     EmotePromptMessage,
+    VariableMessage,
 } from "../Messages/generated/messages_pb";
 import { UserMovesMessage } from "../Messages/generated/messages_pb";
 import { TemplatedApp } from "uWebSockets.js";
@@ -357,6 +358,8 @@ export class IoSocketController {
                     socketManager.handleSilentMessage(client, message.getSilentmessage() as SilentMessage);
                 } else if (message.hasItemeventmessage()) {
                     socketManager.handleItemEvent(client, message.getItemeventmessage() as ItemEventMessage);
+                } else if (message.hasVariablemessage()) {
+                    socketManager.handleVariableEvent(client, message.getVariablemessage() as VariableMessage);
                 } else if (message.hasWebrtcsignaltoservermessage()) {
                     socketManager.emitVideo(
                         client,